requirements.txt
pydantic
groq
fastapi
python-dotenv
instructor[groq]
fastapi[standard]
uvicorn
jinja2